Удалил лишние файлы у vue и разделил на компоненты
